会签加组长

+++ Components/BP.WF/WF/WorkNode.cs
@@ -5328,33 +5328,46 @@ namespace BP.WF
                 string myhuiQianZhuChiRen = gwlOfMe.GetParaString("HuiQianZhuChiRen");
                 string huiQianType = gwlOfMe.GetParaString("huiQianType");
                 //说明是主持人/第二主持人
-                if (this.HisGenerWorkFlow.TodoEmps.Contains(BP.Web.WebUser.No + "," + BP.Web.WebUser.Name + ";") == true && this.HisGenerWorkFlow.GetParaString("AddLeader").Contains(BP.Web.WebUser.No + ",") == false)
+                if (this.HisGenerWorkFlow.TodoEmps.Contains(BP.Web.WebUser.No + "," + BP.Web.WebUser.Name + ";") == true )
                 {
 
                     /*当前人是组长，检查是否可以可以发送,检查自己是否是加签后的最后一个人 ？*/
                     string todoEmps = ""; //记录没有处理的人.
-                    int num = 0;
+                    int num = 0; //主持人自己加签
+                    int leaderNum = 0;
                     foreach (GenerWorkerList item in gwls)
                     {
                         if (item.IsPassInt == 0 || item.IsPassInt == 90)
                         {
                             string huiQianZhuChiRen = item.GetParaString("HuiQianZhuChiRen");
-                            if (item.FK_Emp != WebUser.No && huiQianZhuChiRen.Equals(WebUser.No))
+                            if (item.FK_Emp != WebUser.No && (huiQianZhuChiRen.Equals(WebUser.No) || item.FK_Emp.Equals(myhuiQianZhuChiRen) || huiQianZhuChiRen.Equals(myhuiQianZhuChiRen)))
                                 todoEmps += BP.WF.Glo.DealUserInfoShowModel(item.FK_Emp, item.FK_EmpText) + " ";
-                            //仅只有一个组长
-                            if (this.HisNode.HuiQianLeaderRole == HuiQianLeaderRole.OnlyOne)
-                                num++;
 
-                            //任意组长
-                            if (this.HisNode.HuiQianLeaderRole == HuiQianLeaderRole.EveryOneMain && (huiQianZhuChiRen.Equals(WebUser.No) || item.FK_Emp == WebUser.No))
-                                num++;
+                            if(this.HisGenerWorkFlow.GetParaString("AddLeader").Contains(BP.Web.WebUser.No + ",") == true)
+                            {
+                                if (huiQianZhuChiRen.Equals(WebUser.No) == true || item.FK_Emp == WebUser.No)
+                                    num++;
+                            }
+                            else
+                            {
+                                //仅只有一个组长
+                                if (this.HisNode.HuiQianLeaderRole == HuiQianLeaderRole.OnlyOne)
+                                    num++;
 
-                            //最后一个组长
-                            if (this.HisNode.HuiQianLeaderRole == HuiQianLeaderRole.LastOneMain && (huiQianZhuChiRen.Equals(WebUser.No) || DataType.IsNullOrEmpty(huiQianZhuChiRen) == true))
-                                num++;
+                                //任意组长
+                                if (this.HisNode.HuiQianLeaderRole == HuiQianLeaderRole.EveryOneMain && (huiQianZhuChiRen.Equals(WebUser.No) || item.FK_Emp == WebUser.No))
+                                    num++;
 
-                          
-                               
+                                //最后一个组长
+                                if (this.HisNode.HuiQianLeaderRole == HuiQianLeaderRole.LastOneMain) {
+                                    if(huiQianZhuChiRen.Equals(WebUser.No) || item.FK_Emp == WebUser.No)
+                                        num++;//当前主持人加签的人员及本身
+                                    if (DataType.IsNullOrEmpty(huiQianZhuChiRen) == true)//未通过的主持人
+                                        leaderNum++;
+                                }
+
+                            }
+
                         }
 
                     }
@@ -5362,45 +5375,52 @@ namespace BP.WF
                     /*只有一个待办,说明自己就是最后的一个人.*/
                     if (num == 1)
                     {
-                        this.HisGenerWorkFlow.Sender = BP.WF.Glo.DealUserInfoShowModel(BP.Web.WebUser.No, BP.Web.WebUser.Name);
-                        this.HisGenerWorkFlow.HuiQianTaskSta = HuiQianTaskSta.None;
-
-                        //如果是任意组长可以发送,则需要设置所有的GenerWorkerList待办结束
-
-                        if (this.HisNode.HuiQianLeaderRole == HuiQianLeaderRole.EveryOneMain)
+                        //说明是原始主持人
+                        if (this.HisGenerWorkFlow.GetParaString("AddLeader").Contains(BP.Web.WebUser.No + ",")==false && leaderNum==0)
                         {
-                            string sql = "Delete FROM WF_GenerWorkerList WHERE WorkID=" + this.WorkID + " AND FK_Node=" + this.HisNode.NodeID + " AND IsPass=0";
-                            DBAccess.RunSQL(sql);
+                            this.HisGenerWorkFlow.Sender = BP.WF.Glo.DealUserInfoShowModel(BP.Web.WebUser.No, BP.Web.WebUser.Name);
+                            this.HisGenerWorkFlow.HuiQianTaskSta = HuiQianTaskSta.None;
 
+                            //如果是任意组长可以发送,则需要设置所有的GenerWorkerList待办结束
+
+                            if (this.HisNode.HuiQianLeaderRole == HuiQianLeaderRole.EveryOneMain)
+                            {
+                                string sql = "Delete FROM WF_GenerWorkerList WHERE WorkID=" + this.WorkID + " AND FK_Node=" + this.HisNode.NodeID + " AND IsPass=0";
+                                DBAccess.RunSQL(sql);
+
+                            }
+                            return false;
                         }
-                        return false;
-                    }
-
-                    //把当前的待办设置已办，并且提示未处理的人当前节点是主持人。
-                    foreach (GenerWorkerList gwl in gwls)
-                    {
-                        if (gwl.FK_Emp != WebUser.No)
-                            continue;
-
-                        //设置当前已经完成.
-                        gwl.IsPassInt = 1;
-                        gwl.Update();
-
-                        // 检查完成条件。
-                        if (this.HisNode.IsEndNode == false)
-                            this.CheckCompleteCondition();
-                        //调用发送成功事件.
-                        string sendSuccess = this.HisFlow.DoFlowEventEntity(EventListOfNode.SendSuccess,
-                            this.HisNode, this.rptGe, null, this.HisMsgObjs);
-                        this.HisMsgObjs.AddMsg("info21", sendSuccess, sendSuccess, SendReturnMsgType.Info);
-
-                        //执行时效考核.
-                        if (this.rptGe == null)
-                            Glo.InitCH(this.HisFlow, this.HisNode, this.WorkID, this.rptGe.FID, this.rptGe.Title, gwl);
                         else
-                            Glo.InitCH(this.HisFlow, this.HisNode, this.WorkID, 0, this.HisGenerWorkFlow.Title, gwl);
+                        {
+                            //把当前的待办设置已办，并且提示未处理的人当前节点是主持人。
+                            foreach (GenerWorkerList gwl in gwls)
+                            {
+                                if (gwl.FK_Emp != WebUser.No)
+                                    continue;
 
-                        this.AddToTrack(ActionType.TeampUp, gwl.FK_Emp, todoEmps, this.HisNode.NodeID, this.HisNode.Name, "协作发送");
+                                //设置当前已经完成.
+                                gwl.IsPassInt = 1;
+                                gwl.Update();
+
+                                // 检查完成条件。
+                                if (this.HisNode.IsEndNode == false)
+                                    this.CheckCompleteCondition();
+                                //调用发送成功事件.
+                                string sendSuccess = this.HisFlow.DoFlowEventEntity(EventListOfNode.SendSuccess,
+                                    this.HisNode, this.rptGe, null, this.HisMsgObjs);
+                                this.HisMsgObjs.AddMsg("info21", sendSuccess, sendSuccess, SendReturnMsgType.Info);
+
+                                //执行时效考核.
+                                if (this.rptGe == null)
+                                    Glo.InitCH(this.HisFlow, this.HisNode, this.WorkID, this.rptGe.FID, this.rptGe.Title, gwl);
+                                else
+                                    Glo.InitCH(this.HisFlow, this.HisNode, this.WorkID, 0, this.HisGenerWorkFlow.Title, gwl);
+
+                                this.AddToTrack(ActionType.TeampUp, gwl.FK_Emp, todoEmps, this.HisNode.NodeID, this.HisNode.Name, "协作发送");
+                            }
+                        }
+                       
                     }
 
                     if (SystemConfig.CustomerNo == "LIMS")
